fix i32-i64 for list_functions

+++ core/src/network/test_utils.rs
@@ -98,10 +98,10 @@ pub fn test_wat_always_valid() -> String {
 
     (func
         (export "__list_functions")
-        (param $allocation i32)
-        (result i32)
+        (param $allocation i64)
+        (result i64)
 
-        (i32.const 0)
+        (i64.const 0)
     )
 )
                 "#
@@ -196,10 +196,10 @@ pub fn test_wat_always_invalid() -> String {
 
     (func
         (export "__list_functions")
-        (param $allocation i32)
-        (result i32)
+        (param $allocation i64)
+        (result i64)
 
-        (i32.const 0)
+        (i64.const 0)
     )
 )
                 "#
